v.1.1.2 bug fixed

+++ main1.c
@@ -218,7 +218,7 @@ int  addsans(struct sal x[],int nsalon ,struct mo y[],int nmovie) {
     printf("Inter film's number:");
     scanf("%d", &a);
     printf("Inter time of sans(Year(>2020) Month Day Hour(24) Minute):");
-    long long int ye, mo, d, h, mi;
+    long int ye, mo, d, h, mi;
     scanf("%ld %ld %ld %ld %ld", &ye, &mo, &d, &h, &mi);
     ye -= 2020;
     mo--;
